Fixed build warnings/error on Linux
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@1490 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjlib-util/include/pjlib-util/base64.h
@@ -76,7 +76,7 @@ PJ_DECL(pj_status_t) pj_base64_encode(const pj_uint8_t *input, int in_len,
  *		    length of the output.
  */
 PJ_DECL(pj_status_t) pj_base64_decode(const pj_str_t *input, 
-				      pj_uint8_t *out, long *out_len);
+				      pj_uint8_t *out, int *out_len);
